Fixed issue with notices when previewing a question dev Fixed other minor issues with template editor New feature: Tabs can be inserted now in template editing area
git-svn-id: file:///Users/Shitiz/Downloads/lssvn/source/limesurvey@6913 b72ed6b6-b9f8-46b5-92b4-906544132732

+++ common.php
@@ -6293,7 +6293,7 @@ function cleanTempDirectory()
     $dir=  $tempdir.'/';
     $dp = opendir($dir) or die ('Could not open temporary directory');
     while ($file = readdir($dp)) {
-        if ((filemtime($dir.$file)) < (strtotime('-1 days')) && $file!='index.html' && $file!='readme.txt' && $file!='..' && $file!='.') {
+        if ((filemtime($dir.$file)) < (strtotime('-1 days')) && $file!='index.html' && $file!='readme.txt' && $file!='..' && $file!='.' && $file!='.svn') {
             unlink($dir.$file);
         }
     }
